Uploaded SQL files for pizza conversion project
Added SQL scripts that are part of a project to create a process to convert the format of a pizza shop's collection of order files into something easily imported and converted using a BULK INSERT and stored procedures. These scripts are just prototypes, the final versions will be uploaded shortly.